+++ templates/init/main.lua
@@ -1,0 +1,10 @@
+function _init()
+end
+
+function _update(dt)
+end
+
+function _draw(dt)
+  gfx.clear(gfx.COLOR_BLACK)
+  gfx.text("Hello, Usagi!", 10, 10, gfx.COLOR_WHITE)
+end